% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covshrink.R
\docType{package}
\name{covshrink}
\alias{covshrink}
\title{covshrink: The package computes the shrinked covariance matrix
proposed by Ledoit and Wolfe (2003).

Currently the package exports only a single function:
cov_shrink}
\section{covshrink functions}{

The foo functions ...
}

