% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covshrink.R
\docType{package}
\name{foo}
\alias{foo}
\title{covshrink: A package for computing the shrinked covariance matrix
proposed by Ledoit and Wolf

The foo package provides three categories of important functions:
foo, bar and baz.}
\section{Foo functions}{

The foo functions ...
}

