% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_hat.R
\name{gamma_hat}
\alias{gamma_hat}
\title{Estimate of gamma}
\usage{
gamma_hat(mat_cov, mat_structurded, na.rm)
}
\arguments{
\item{mat_cov}{matrix containing sample covariance matrix.}

\item{mat_structurded}{matrix containing structured covariance matrix.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped
before the computation proceeds.}
}
\value{
matrix of shape \eqn{p \times p}.
}
\description{
Estimate of gamma
}
