% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_obs.R
\name{pairwise_obs}
\alias{pairwise_obs}
\title{Compute the number of pairwise complete observations}
\usage{
pairwise_obs(data)
}
\arguments{
\item{data}{dataframe or matrix}
}
\value{
matrix of integers giving the number of pairwise complete observations.
}
\description{
Compute the number of pairwise complete observations
}
