% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_nu_hat.R
\name{pi_nu_hat}
\alias{pi_nu_hat}
\title{Estimator of pi_ij and nu_ij}
\usage{
pi_nu_hat(data, mat_cov, mat_obs, na.rm = FALSE)
}
\arguments{
\item{data}{matrix or data.frame containing only numeric variables}

\item{mat_cov}{matrix containing sample covariance estimates.}

\item{mat_obs}{matrix of pairwise observations.}

\item{na.rm}{logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}
}
\value{
matrix of individual estimates for \eqn{\pi_{ij}} and the entries
\eqn{\sqrt{s_{j}^2 / s_{i}^2} \times \nu_{ii, ij}} and
\eqn{\sqrt{s_{i}^2 / s_{j}^2} \times \nu_{jj, ij}}.
}
\description{
Estimator of pi_ij and nu_ij
}
