% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_wise_full_product}
\alias{row_wise_full_product}
\title{Title}
\usage{
row_wise_full_product(data)
}
\arguments{
\item{data}{Dataframe or matrix of numeric values}
}
\value{
Array of rowwise full products
}
\description{
Title
}
