% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_cov.R
\name{structured_cov}
\alias{structured_cov}
\title{Estimate of the structured covariance matrix}
\usage{
structured_cov(mat_covariance, na.rm = FALSE, mat_obs_weight = NULL)
}
\arguments{
\item{mat_covariance}{matrix containing sample covariance estimates.}

\item{na.rm}{logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}

\item{mat_obs_weight}{matrix of pairwise observations. If provided an
observations weighted average of the correlation coefficients is
computed giving higher weight to coefficients computed from more data.}
}
\value{
matrix containing the \eqn{\bar{\rho} \times s_{i} \times s_{j}} on
for variable \eqn{i} and \eqn{j} on the nondiagonal and \eqn{s_i^2} on
the diagonal.
}
\description{
Estimate of the structured covariance matrix
}
